+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/PlatformProxy.java
@@ -14,15 +14,19 @@ import com.refinedmods.refinedstorage2.platform.common.packet.ServerToClientComm
 import com.refinedmods.refinedstorage2.platform.common.render.FluidRenderer;
 import com.refinedmods.refinedstorage2.platform.common.util.BucketAmountFormatting;
 
+import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 import javax.annotation.Nullable;
 
 import net.minecraft.client.KeyMapping;
+import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.EditBox;
+import net.minecraft.client.gui.screens.inventory.tooltip.ClientTooltipComponent;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.core.NonNullList;
+import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.sounds.SoundEvent;
 import net.minecraft.tags.TagKey;
@@ -30,6 +34,7 @@ import net.minecraft.world.Container;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.inventory.CraftingContainer;
+import net.minecraft.world.inventory.tooltip.TooltipComponent;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.CraftingRecipe;
@@ -205,6 +210,23 @@ public class PlatformProxy implements Platform {
     @Override
     public Optional<SoundEvent> getBucketPickupSound(final LiquidBlock liquidBlock, final BlockState state) {
         return ensureLoaded().getBucketPickupSound(liquidBlock, state);
+    }
+
+    @Override
+    public List<ClientTooltipComponent> processTooltipComponents(final ItemStack stack,
+                                                                 final GuiGraphics graphics,
+                                                                 final int mouseX,
+                                                                 final Optional<TooltipComponent> imageComponent,
+                                                                 final List<Component> components) {
+        return ensureLoaded().processTooltipComponents(stack, graphics, mouseX, imageComponent, components);
+    }
+
+    @Override
+    public void renderTooltip(final GuiGraphics graphics,
+                              final List<ClientTooltipComponent> components,
+                              final int x,
+                              final int y) {
+        ensureLoaded().renderTooltip(graphics, components, x, y);
     }
 
     private Platform ensureLoaded() {