+++ CHANGELOG.md
@@ -20,6 +20,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 -   Fixed Grid voiding fluids if there was no space in inventory on Fabric.
 -   Fixed Grid dropping fluid buckets if there was no space in inventory on Forge.
+-   Fixed compatibility with custom tooltips in the Grid.
+-   Fixed bundle tooltip in the Grid.
 
 ## [2.0.0-milestone.2.11] - 2023-07-04